Skip tui askvnc reboot for dirinstall (#1164254)
When running with --dirinstall, or image or live we don't want to reboot
the system when quitting tui mode. Normally this is handled by the
exitHandler, but askvnc is too early for that to work.

Resolves: rhbz#1164254

+++ pyanaconda/ui/tui/spokes/askvnc.py
@@ -22,12 +22,14 @@
 from pyanaconda.ui.tui.spokes import NormalTUISpoke
 from pyanaconda.ui.tui.simpleline import TextWidget, ColumnWidget
 from pyanaconda.ui.tui.tuiobject import YesNoDialog
-from pyanaconda.constants import USEVNC, USETEXT
+from pyanaconda.constants import USEVNC, USETEXT, IPMI_ABORTED
 from pyanaconda.constants_text import INPUT_PROCESSED
 from pyanaconda.i18n import N_, _
 from pyanaconda.ui.communication import hubQ
 from pyanaconda.ui.tui import exception_msg_handler
 from pyanaconda.iutil import execWithRedirect
+from pyanaconda.flags import can_touch_runtime_system
+from pyanaconda import iutil
 import getpass
 import sys
 
@@ -105,7 +107,11 @@ class AskVNCSpoke(NormalTUISpoke):
             d = YesNoDialog(self.app, _(self.app.quit_message))
             self.app.switch_screen_modal(d)
             if d.answer:
-                execWithRedirect("systemctl", ["--no-wall", "reboot"])
+                iutil.ipmi_report(IPMI_ABORTED)
+                if can_touch_runtime_system("Quit and Reboot"):
+                    execWithRedirect("systemctl", ["--no-wall", "reboot"])
+                else:
+                    exit(1)
         else:
             return key
